MorphoRegions: Analysis of Regionalization Patterns in Serially-Homologous Structures

Computes the optimal number of regions (or subdivisions) and their position in serial structures without a priori assumptions and to visualize the results. After reducing data dimensionality with the built-in function for data ordination, regions are fitted as segmented linear regressions along the serial structure. Every region boundary position and increasing number of regions are iteratively fitted and the best model (number of regions and boundary positions) is selected with an information criterion. This package expands on the previous 'regions' package (Jones et al., Science 2018) with improved computation and more fitting and plotting options.

See also

Author

Maintainer: Amandine Gillet gillet.aman@gmail.com (ORCID)

Authors:

Katrina Jones katrina.jones@manchester.ac.uk (ORCID)

Noah Greifer ngreifer@iq.harvard.edu (ORCID)

Stephanie Pierce spierce@oeb.harvard.edu (ORCID)